Move MemberSelect under Member module.
